@@ -11,63 +11,134 @@
  * @link       https://itea3.org
  */
 
+declare(strict_types=1);
+
 namespace Calendar\Controller\Plugin;
 
 use Calendar\Entity\Calendar;
 use Calendar\Entity\Contact as CalendarContact;
 use Calendar\Options\ModuleOptions;
 use Calendar\Service\CalendarService;
-use General\Service\GeneralService;
 use Zend\Mvc\Controller\Plugin\AbstractPlugin;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use ZfcTwig\View\TwigRenderer;
 
 /**
- * Create a link to an project.
+ * Class RenderCalendarContactList
  *
- * @category   Program
- *
- * @author     Johan van der Heide <johan.van.der.heide@itea3.org>
- * @license    https://itea3.org/licence.txt proprietary
- *
- * @link       https://itea3.org
+ * @package Calendar\Controller\Plugin
  */
 class RenderCalendarContactList extends AbstractPlugin
 {
     /**
-     * @var ServiceLocatorInterface
+     * @var TwigRenderer
      */
-    protected $serviceLocator;
+    protected $twigRenderer;
+    /**
+     * @var ModuleOptions
+     */
+    protected $moduleOptions;
+    /**
+     * @var CalendarService
+     */
+    protected $calendarService;
+
+    /**
+     * RenderCalendarContactList constructor.
+     *
+     * @param TwigRenderer    $twigRenderer
+     * @param ModuleOptions   $moduleOptions
+     * @param CalendarService $calendarService
+     */
+    public function __construct(
+        TwigRenderer $twigRenderer,
+        ModuleOptions $moduleOptions,
+        CalendarService $calendarService
+    ) {
+        $this->twigRenderer = $twigRenderer;
+        $this->moduleOptions = $moduleOptions;
+        $this->calendarService = $calendarService;
+    }
+
 
     /**
      * @param Calendar $calendar
      *
      * @return CalendarPdf
      */
-    public function render(Calendar $calendar)
+    public function renderPresenceList(Calendar $calendar): CalendarPdf
     {
         $pdf = new CalendarPdf();
-        $pdf->setTemplate($this->getModuleOptions()->getCalendarContactTemplate());
+        $pdf->setTemplate($this->moduleOptions->getCalendarContactTemplate());
         $pdf->AddPage();
-        $pdf->SetFontSize(9);
-        $twig = $this->getServiceLocator()->get('ZfcTwigRenderer');
+        $pdf->SetFont('freesans', '', 10);
 
-        $calendarContacts = $this->getCalendarService()
-                                 ->findCalendarContactsByCalendar($calendar, CalendarContact::STATUS_NO_DECLINED);
+        $calendarContacts = $this->calendarService->findCalendarContactsByCalendar(
+            $calendar,
+            CalendarContact::STATUS_ALL,
+            'organisation'
+        );
+
+        //Create chunks of arrays per 22, as that amount fits on the screen
+        $paginatedContacts = array_chunk($calendarContacts, 22);
+        $minAmountOfPages = max(\count($paginatedContacts), 1);
+
+        for ($i = 0; $i < $minAmountOfPages; $i++) {
+            /**
+             *
+             */
+            $contactListContent = $this->twigRenderer->render(
+                'calendar/pdf/presence-list',
+                [
+                    'calendarService'  => $this->calendarService,
+                    'calendar'         => $calendar,
+                    'calendarContacts' => $paginatedContacts[$i] ?? [],
+                ]
+            );
+
+            $pdf->writeHTMLCell(0, 0, 14, 42, $contactListContent);
+
+            /*
+            * Don't add a new page on the last iteration
+            */
+            if ($i < $minAmountOfPages - 1) {
+                $pdf->AddPage();
+            }
+        }
+
+        return $pdf;
+    }
+
+    /**
+     * @param Calendar $calendar
+     *
+     * @return CalendarPdf
+     */
+    public function renderSignatureList(Calendar $calendar): CalendarPdf
+    {
+        $pdf = new CalendarPdf();
+        $pdf->setTemplate($this->moduleOptions->getCalendarContactTemplate());
+        $pdf->AddPage();
+        $pdf->SetFont('freesans', '', 10);
+
+        $calendarContacts = $this->calendarService->findCalendarContactsByCalendar(
+            $calendar,
+            CalendarContact::STATUS_NO_DECLINED
+        );
 
         //Create chunks of arrays per 13, as that amount fits on the screen
         $paginatedContacts = array_chunk($calendarContacts, 13);
-        $minAmountOfPages  = max(count($paginatedContacts), 2);
+        $minAmountOfPages = max(\count($paginatedContacts), 2);
 
         for ($i = 0; $i < $minAmountOfPages; $i++) {
             /*
              * Use the NDA object to render the filename
              */
-            $contactListContent = $twig->render(
-                'calendar/pdf/calendar-contact',
+            $contactListContent = $this->twigRenderer->render(
+                'calendar/pdf/signature-list',
                 [
-                    'calendarService'  => $this->getCalendarService(),
+                    'calendarService'  => $this->calendarService,
                     'calendar'         => $calendar,
-                    'calendarContacts' => isset($paginatedContacts[$i]) ? $paginatedContacts[$i] : [],
+                    'calendarContacts' => $paginatedContacts[$i] ?? [],
                 ]
             );
 
@@ -82,51 +153,5 @@ class RenderCalendarContactList extends AbstractPlugin
         }
 
         return $pdf;
-    }
-
-    /**
-     * @return ModuleOptions
-     */
-    public function getModuleOptions()
-    {
-        return $this->getServiceLocator()->get(ModuleOptions::class);
-    }
-
-    /**
-     * @return ServiceLocatorInterface
-     */
-    public function getServiceLocator()
-    {
-        return $this->serviceLocator;
-    }
-
-    /**
-     * @param ServiceLocatorInterface $serviceLocator
-     *
-     * @return $this
-     */
-    public function setServiceLocator(ServiceLocatorInterface $serviceLocator)
-    {
-        $this->serviceLocator = $serviceLocator;
-
-        return $this;
-    }
-
-    /**
-     * @return CalendarService
-     */
-    public function getCalendarService()
-    {
-        return $this->getServiceLocator()->get(CalendarService::class);
-    }
-
-    /**
-     * Gateway to the General Service.
-     *
-     * @return GeneralService
-     */
-    public function getGeneralService()
-    {
-        return $this->getServiceLocator()->get(GeneralService::class);
     }
 }
