@@ -1,17 +1,71 @@
 <?php
-_("txt-final");
-_("txt-calendar-sequence-explanation");
-_("txt-date-from");
-_("txt-date-from-explanation");
-_("txt-date-end");
-_("txt-date-end-explanation");
-_("txt-calendar-description-explanation");
-_("txt-image-url-explanation");
-_("txt-calendar-url-explanation");
-_("txt-calendar-explanation");
+_("txt-calendar-calendar-placeholder");
+_("txt-calendar-calendar-label");
+_("txt-calendar-calendar-help-block");
+_("txt-calendar-location-placeholder");
+_("txt-calendar-location-label");
+_("txt-calendar-location-help-block");
+_("txt-calendar-date-from-label");
+_("txt-calendar-date-from-help-block");
+_("txt-calendar-date-end-label");
+_("txt-calendar-date-end-help-block");
+_("txt-calendar-final-label");
+_("txt-calendar-final-help-block");
+_("txt-calendar-on-homepage-label");
+_("txt-calendar-on-homepage-help-block");
+_("txt-calendar-highlight-label");
+_("txt-calendar-highlight-help-block");
+_("txt-calendar-highlight-description-placeholder");
+_("txt-calendar-highlight-description-label");
+_("txt-calendar-highlight-description-help-block");
+_("txt-calendar-own-event-label");
+_("txt-calendar-own-event-help-block");
+_("txt-calendar-present-label");
+_("txt-calendar-present-help-block");
+_("txt-calendar-sequence-label");
+_("txt-calendar-sequence-help-block");
+_("txt-calendar-description-placeholder");
+_("txt-calendar-description-label");
+_("txt-calendar-description-help-block");
+_("txt-calendar-url-placeholder");
+_("txt-calendar-url-label");
+_("txt-calendar-url-help-block");
+_("txt-calendar-image-url-placeholder");
+_("txt-calendar-image-url-label");
+_("txt-calendar-image-url-help-block");
+_("txt-calendar-type-label");
+_("txt-calendar-type-help-block");
+_("txt-program-call");
+_("txt-calendar-program-call-help-block");
 _("txt-draft");
 _("txt-tentative");
-_("txt-location-explanation");
-
-_("txt-final-explanation");
-_("txt-on-homepage-explanation");
+_("txt-final");
+_("txt-not-on-homepage");
+_("txt-on-homepage");
+_("txt-no-highlight");
+_("txt-highlight-on-event-page");
+_("txt-no-own-event");
+_("txt-own-event");
+_("txt-office-not-present");
+_("txt-office-present");
+_("txt-calendar-contact-role-role-placeholder");
+_("txt-calendar-contact-role-role-label");
+_("txt-calendar-contact-role-role-help-block");
+_("txt-role-attendees");
+_("txt-stg-reviewer");
+_("txt-stg-spare-reviewer");
+_("txt-calendar-contact-status-status-placeholder");
+_("txt-calendar-contact-status-status-label");
+_("txt-calendar-contact-status-status-help-block");
+_("txt-calendar-contact-status-status-change-placeholder");
+_("txt-calendar-contact-status-status-change-label");
+_("txt-calendar-contact-status-status-change-help-block");
+_("txt-calendar-type-type-placeholder");
+_("txt-calendar-type-type-label");
+_("txt-calendar-type-type-help-block");
+_("txt-calendar-type-background-color-label");
+_("txt-calendar-type-background-color-help-block");
+_("txt-calendar-font-color-label");
+_("txt-calendar-font-color-help-block");
+_("txt-calendar-type-access-label");
+_("txt-calendar-type-access-help-block");
