@@ -8,6 +8,8 @@
  * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
  */
 
+declare(strict_types=1);
+
 namespace Calendar\Entity;
 
 use Doctrine\Common\Collections;
@@ -18,21 +20,38 @@ use Zend\Form\Annotation;
  * CalendarContactRole.
  *
  * @ORM\Table(name="calendar_contact_role")
- * @ORM\Entity
+ * @ORM\Entity(repositoryClass="Calendar\Repository\ContactRole")
  */
-class ContactRole
+class ContactRole extends AbstractEntity
 {
-    const ROLE_ATTENDEE = 1;
+    public const ROLE_ATTENDEE = 1;
+    public const ROLE_STG_REVIEWER = 7;
+    public const ROLE_STG_SPARE_REVIEWER = 8;
+
+    /**
+     * @var array Lookup table for the roles
+     */
+    public static $roles
+        = [
+            self::ROLE_ATTENDEE           => 'txt-role-attendees',
+            self::ROLE_STG_REVIEWER       => 'txt-stg-reviewer',
+            self::ROLE_STG_SPARE_REVIEWER => 'txt-stg-spare-reviewer',
+        ];
+
     /**
      * @ORM\Column(name="role_id", type="integer", nullable=false)
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
+     * @Annotation\Type("\Zend\Form\Element\Hidden")
      *
      * @var integer
      */
     private $id;
     /**
      * @ORM\Column(name="role", type="string", length=45, nullable=false)
+     * @Annotation\Type("\Zend\Form\Element\Text")
+     * @Annotation\Options({"label":"txt-calendar-contact-role-role-label","help-block": "txt-calendar-contact-role-role-help-block"})
+     * @Annotation\Attributes({"placeholder":"txt-calendar-contact-role-role-placeholder"})     *
      *
      * @var string
      */
@@ -45,27 +64,31 @@ class ContactRole
      */
     private $calendarContact;
 
-    /**
-     * Class constructor.
-     */
     public function __construct()
     {
         $this->calendarContact = new Collections\ArrayCollection();
     }
 
-    /**
-     * Return the name of the role.
-     *
-     * @return string
-     */
-    public function __toString()
+    public function __get($property)
+    {
+        return $this->$property;
+    }
+
+    public function __set($property, $value)
+    {
+        $this->$property = $value;
+    }
+
+    public function __isset($property)
+    {
+        return isset($this->$property);
+    }
+
+    public function __toString(): string
     {
         return (string)$this->role;
     }
 
-    /**
-     * @return int
-     */
     public function getId()
     {
         return $this->id;
@@ -76,10 +99,9 @@ class ContactRole
      *
      * @return ContactRole
      */
-    public function setId($id)
+    public function setId($id): ContactRole
     {
         $this->id = $id;
-
         return $this;
     }
 
@@ -96,10 +118,9 @@ class ContactRole
      *
      * @return ContactRole
      */
-    public function setRole($role)
+    public function setRole(string $role): ContactRole
     {
         $this->role = $role;
-
         return $this;
     }
 
@@ -116,10 +137,9 @@ class ContactRole
      *
      * @return ContactRole
      */
-    public function setCalendarContact($calendarContact)
+    public function setCalendarContact(array $calendarContact): ContactRole
     {
         $this->calendarContact = $calendarContact;
-
         return $this;
     }
 }
