@@ -12,6 +12,8 @@
  *
  * @link        http://github.com/iteaoffice/main for the canonical source repository
  */
+declare(strict_types=1);
+
 namespace Calendar\Factory;
 
 use Calendar\Service\FormService;
@@ -26,24 +28,9 @@ use Zend\ServiceManager\Factory\FactoryInterface;
  */
 final class FormServiceFactory implements FactoryInterface
 {
-    /**
-     * Create an instance of the requested class name.
-     *
-     * @param ContainerInterface $container
-     * @param string             $requestedName
-     * @param null|array         $options
-     *
-     * @return FormService
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): FormService
     {
         /** @var FormService $formService */
-        $formService = new $requestedName($options);
-        $formService->setServiceLocator($container);
-        /** @var EntityManager $entityManager */
-        $entityManager = $container->get(EntityManager::class);
-        $formService->setEntityManager($entityManager);
-
-        return $formService;
+        return new $requestedName($container, $container->get(EntityManager::class));
     }
 }
