@@ -13,62 +13,24 @@
  * @link        http://github.com/iteaoffice/project for the canonical source repository
  */
 
+declare(strict_types=1);
+
 namespace Calendar;
 
-use Calendar\Controller\Plugin\RenderCalendarContactList;
-use Calendar\Controller\Plugin\RenderReviewCalendar;
 use Calendar\Navigation;
 use Zend\ModuleManager\Feature;
-use Zend\Mvc\Controller\PluginManager;
 
 /**
  * @author
  */
-class Module implements Feature\AutoloaderProviderInterface, Feature\ConfigProviderInterface
+class Module implements Feature\ConfigProviderInterface
 {
-    /**
-     * @return array
-     */
-    public function getAutoloaderConfig()
-    {
-        return [
-            'Zend\Loader\ClassMapAutoloader' => [
-                __DIR__ . '/../autoload_classmap.php',
-            ],
-        ];
-    }
 
     /**
-     * @return array
+     * F * @return array
      */
-    public function getConfig()
+    public function getConfig(): array
     {
         return include __DIR__ . '/../config/module.config.php';
-    }
-
-    /**
-     * Move this to here to have config cache working.
-     *
-     * @return array
-     */
-    public function getControllerPluginConfig()
-    {
-        return [
-            'factories' => [
-                'renderCalendarContactList' => function (PluginManager $sm) {
-                    $renderCalendarContactList
-                        = new RenderCalendarContactList();
-                    $renderCalendarContactList->setServiceLocator($sm->getServiceLocator());
-
-                    return $renderCalendarContactList;
-                },
-                'renderReviewCalendar'      => function (PluginManager $sm) {
-                    $renderReviewCalendar = new RenderReviewCalendar();
-                    $renderReviewCalendar->setServiceLocator($sm->getServiceLocator());
-
-                    return $renderReviewCalendar;
-                },
-            ],
-        ];
     }
 }
