@@ -7,41 +7,49 @@
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
  * @copyright Copyright (calendar_entity_calendar) Copyright (c) 2004-2017 ITEA Office (https://itea3.org) (https://itea3.org)
  */
+declare(strict_types=1);
+
 namespace Calendar\Repository;
 
 use Admin\Entity\Access;
 use Calendar\Entity;
 use Calendar\Service\CalendarService;
 use Contact\Entity\Contact;
+use Doctrine\Common\Collections\Criteria;
 use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
+use DoctrineExtensions\Query\Mysql\Year;
 use Project\Entity\Project;
 
 /**
- * @category    Calendar
- * @package     Repository
+ * Class Calendar
+ *
+ * @package Calendar\Repository
  */
-class Calendar extends EntityRepository
+final class Calendar extends EntityRepository
 {
-    /**
-     * @param              $which
-     * @param bool         $filterForAccess
-     * @param Contact|null $contact
-     * @param null         $year
-     * @param null         $type
-     *
-     * @return \Doctrine\ORM\Query
-     */
     public function findCalendarItems(
-        $which,
-        $filterForAccess = true,
+        string $which,
+        bool $filterForAccess = true,
         Contact $contact = null,
-        $year = null,
-        $type = null
-    ) {
+        int $year = null,
+        string $type = null,
+        ?QueryBuilder $limitQueryBuilder = null
+    ): Query {
         $qb = $this->_em->createQueryBuilder();
         $qb->select('calendar_entity_calendar');
         $qb->from(Entity\Calendar::class, 'calendar_entity_calendar');
+
+        //Add a sub-select to be able to filter on public
+        $subSelect = $this->_em->createQueryBuilder();
+        $subSelect->select('calendar_entity_calendar_type');
+        $subSelect->from(Entity\Type::class, 'calendar_entity_calendar_type');
+        $subSelect->join('calendar_entity_calendar_type.access', 'admin_entity_access');
+        $subSelect->andWhere(
+            $qb->expr()
+                ->in('admin_entity_access.access', [Access::ACCESS_PUBLIC])
+        );
 
 
         switch ($which) {
@@ -53,13 +61,13 @@ class Calendar extends EntityRepository
                 $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
                 break;
             case CalendarService::WHICH_PAST:
-                $qb->andWhere('calendar_entity_calendar.dateEnd < ?1');
+                $qb->andWhere('calendar_entity_calendar.dateFrom < ?1');
 
                 if (null !== $type) {
                     $qb->andWhere('calendar_entity_calendar.type = ?9');
                     $qb->setParameter(9, $type);
                 }
-                $qb->orderBy('calendar_entity_calendar.dateEnd', 'DESC');
+                $qb->orderBy('calendar_entity_calendar.dateFrom', 'DESC');
                 $qb->setParameter(1, new \DateTime());
                 break;
             case CalendarService::WHICH_REVIEWS:
@@ -68,21 +76,23 @@ class Calendar extends EntityRepository
                 $qb->setParameter(1, new \DateTime());
                 $qb->andWhere('calendar_entity_calendar.final = ?3');
                 $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
+
                 $projectCalendarSubSelect = $this->_em->createQueryBuilder();
                 $projectCalendarSubSelect->select('calendar.id');
-                $projectCalendarSubSelect->from('Project\Entity\Calendar\Calendar', 'projectCalendar');
+                $projectCalendarSubSelect->from(\Project\Entity\Calendar\Calendar::class, 'projectCalendar');
                 $projectCalendarSubSelect->join('projectCalendar.calendar', 'calendar');
                 $qb->andWhere($qb->expr()->in('calendar_entity_calendar.id', $projectCalendarSubSelect->getDQL()));
                 break;
             case CalendarService::WHICH_FINAL:
                 $qb->andWhere('calendar_entity_calendar.final = ?3');
                 $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
-                $qb->orderBy('calendar_entity_calendar.dateFrom', 'ASC');
-                $qb->addOrderBy('calendar_entity_calendar.sequence', 'ASC');
+                $qb->orderBy('calendar_entity_calendar.dateFrom', Criteria::DESC);
+                $qb->addOrderBy('calendar_entity_calendar.sequence', Criteria::ASC);
                 break;
             case CalendarService::WHICH_UPDATED:
-                $qb->orderBy('calendar_entity_calendar.dateUpdated', 'DESC');
+                $qb->orderBy('calendar_entity_calendar.dateUpdated', Criteria::DESC);
                 $qb->andWhere('calendar_entity_calendar.final = ?3');
+                $qb->andWhere($qb->expr()->isNotNull('calendar_entity_calendar.dateUpdated'));
                 $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
                 break;
             case CalendarService::WHICH_ON_HOMEPAGE:
@@ -92,29 +102,51 @@ class Calendar extends EntityRepository
                 $qb->setParameter(2, Entity\Calendar::ON_HOMEPAGE);
                 $qb->andWhere('calendar_entity_calendar.final = ?3');
                 $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
-                $qb->orderBy('calendar_entity_calendar.sequence', 'ASC');
-                $qb->addOrderBy('calendar_entity_calendar.dateFrom', 'ASC');
+                $qb->addOrderBy('calendar_entity_calendar.dateFrom', Criteria::ASC);
+                $qb->orderBy('calendar_entity_calendar.sequence', Criteria::ASC);
+
+                //We only want public events
+                $qb->andWhere(
+                    $qb->expr()->in('calendar_entity_calendar.type', $subSelect->getDQL())
+                );
+
+                break;
+            case CalendarService::WHICH_HIGHLIGHT:
+                $qb->andWhere('calendar_entity_calendar.dateEnd >= ?1');
+                $qb->setParameter(1, new \DateTime());
+                $qb->andWhere('calendar_entity_calendar.onHomepage = ?2');
+                $qb->setParameter(2, Entity\Calendar::ON_HOMEPAGE);
+                $qb->andWhere('calendar_entity_calendar.final = ?3');
+                $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
+                $qb->andWhere('calendar_entity_calendar.highlight = ?3');
+                $qb->setParameter(3, Entity\Calendar::HIGHLIGHT);
+                $qb->orderBy('calendar_entity_calendar.sequence', Criteria::ASC);
+                $qb->addOrderBy('calendar_entity_calendar.dateFrom', Criteria::ASC);
+
+                //We only want public events
+                $qb->andWhere(
+                    $qb->expr()->in('calendar_entity_calendar.type', $subSelect->getDQL())
+                );
                 break;
         }
-
 
         if ($filterForAccess) {
             /**
              * When no contact is given, simply return all the public calendar items
              */
-            if (is_null($contact)) {
+            if (null === $contact) {
                 $contact = new Contact();
                 $contact->setId(0);
                 $access = new Access();
                 $access->setAccess(Access::ACCESS_PUBLIC);
                 $contact->setAccess([$access]);
             }
-            $qb = $this->filterForAccess($qb, $contact);
+            $qb = $this->filterForAccess($qb, $contact, $limitQueryBuilder);
         }
 
-        if (! is_null($year)) {
+        if (null !== $year) {
             $emConfig = $this->getEntityManager()->getConfiguration();
-            $emConfig->addCustomDatetimeFunction('YEAR', 'DoctrineExtensions\Query\Mysql\Year');
+            $emConfig->addCustomDatetimeFunction('YEAR', Year::class);
             $qb->andWhere('YEAR(calendar_entity_calendar.dateEnd) = ?8');
             $qb->setParameter(8, (int)$year);
         }
@@ -122,14 +154,11 @@ class Calendar extends EntityRepository
         return $qb->getQuery();
     }
 
-    /**
-     * @param QueryBuilder $qb
-     * @param Contact      $contact
-     *
-     * @return QueryBuilder $qb
-     */
-    public function filterForAccess(QueryBuilder $qb, Contact $contact)
-    {
+    public function filterForAccess(
+        QueryBuilder $qb,
+        Contact $contact,
+        ?QueryBuilder $limitQueryBuilder = null
+    ): QueryBuilder {
         //Filter based on the type access type
         $subSelect = $this->_em->createQueryBuilder();
         $subSelect->select('type');
@@ -137,35 +166,32 @@ class Calendar extends EntityRepository
         $subSelect->join('type.access', 'access');
         $subSelect->andWhere(
             $qb->expr()
-               ->in('access.access', array_merge([strtolower(Access::ACCESS_PUBLIC)], $contact->getRoles()))
+                ->in('access.access', array_merge([strtolower(Access::ACCESS_PUBLIC)], $contact->getRoles()))
         );
 
-        $subSelectCalendarContact = $this->_em->createQueryBuilder();
-        $subSelectCalendarContact->select('calendar2');
-        $subSelectCalendarContact->from('Calendar\Entity\Calendar', 'calendar2');
-        $subSelectCalendarContact->join('calendar2.calendarContact', 'calenderContact2');
-        $subSelectCalendarContact->join('calenderContact2.contact', 'contact2');
-        $subSelectCalendarContact->andWhere('contact2.id = ' . $contact);
-
-        $qb->andWhere(
-            $qb->expr()->orX(
-                $qb->expr()->in(
-                    'calendar_entity_calendar.type',
-                    $subSelect->getDQL()
-                ),
-                $qb->expr()->in('calendar_entity_calendar', $subSelectCalendarContact->getDQL())
-            )
-        );
+        /**
+         * When the permit gives no result, do nothing
+         */
+        if (null === $limitQueryBuilder) {
+            $qb->andWhere(
+                $qb->expr()->in('calendar_entity_calendar.type', $subSelect->getDQL())
+            );
+        } else {
+            $qb->andWhere(
+                $qb->expr()->orX(
+                    $qb->expr()->in(
+                        'calendar_entity_calendar.type',
+                        $subSelect->getDQL()
+                    ),
+                    $qb->expr()->in('calendar_entity_calendar', $limitQueryBuilder->getDQL())
+                )
+            );
+        }
 
         return $qb;
     }
 
-    /**
-     * @param Project $project
-     *
-     * @return Entity\Calendar|null
-     */
-    public function findLatestProjectCalendar(Project $project)
+    public function findLatestProjectCalendar(Project $project): ?Entity\Calendar
     {
         $qb = $this->_em->createQueryBuilder();
         $qb->select('calendar_entity_calendar');
@@ -185,33 +211,7 @@ class Calendar extends EntityRepository
         return $qb->getQuery()->getOneOrNullResult();
     }
 
-    /**
-     * @return array
-     */
-    public function findMinAndMaxYear()
-    {
-        $emConfig = $this->getEntityManager()->getConfiguration();
-        $emConfig->addCustomDatetimeFunction('YEAR', 'DoctrineExtensions\Query\Mysql\Year');
-
-        $dql
-            = 'SELECT 
-                        MIN(YEAR(calendar_entity_calendar.dateFrom)) AS minYear,
-                        MAX(YEAR(calendar_entity_calendar.dateFrom)) AS maxYear
-                   FROM Calendar\Entity\Calendar calendar_entity_calendar';
-
-
-        $result = $this->_em->createQuery($dql)->getScalarResult();
-
-        return array_shift($result);
-    }
-
-    /**
-     * @param Project   $project
-     * @param \DateTime $dateTime
-     *
-     * @return Entity\Calendar|null
-     */
-    public function findNextProjectCalendar(Project $project, \DateTime $dateTime)
+    public function findNextProjectCalendar(Project $project, \DateTime $dateTime): ?Entity\Calendar
     {
         $qb = $this->_em->createQueryBuilder();
         $qb->select('calendar_entity_calendar');
@@ -232,13 +232,7 @@ class Calendar extends EntityRepository
         return $qb->getQuery()->getOneOrNullResult();
     }
 
-    /**
-     * @param Project   $project
-     * @param \DateTime $dateTime
-     *
-     * @return Entity\Calendar|null
-     */
-    public function findPreviousProjectCalendar(Project $project, \DateTime $dateTime)
+    public function findPreviousProjectCalendar(Project $project, \DateTime $dateTime): ?Entity\Calendar
     {
         $qb = $this->_em->createQueryBuilder();
         $qb->select('calendar_entity_calendar');
@@ -259,21 +253,13 @@ class Calendar extends EntityRepository
         return $qb->getQuery()->getOneOrNullResult();
     }
 
-    /**
-     * Function which returns true/false based ont he fact if a user can view the calendar
-     *
-     * @param Entity\Calendar $calendar
-     * @param Contact         $contact
-     *
-     * @return bool
-     */
-    public function canViewCalendar(Entity\Calendar $calendar, Contact $contact = null)
+    public function canViewCalendar(Entity\Calendar $calendar, Contact $contact = null): bool
     {
         $qb = $this->_em->createQueryBuilder();
         $qb->select('calendar_entity_calendar');
         $qb->from(Entity\Calendar::class, 'calendar_entity_calendar');
 
-        if ($contact->isEmpty()) {
+        if (null === $contact) {
             $contact = new Contact();
             $contact->setId(0);
             $access = new Access();
@@ -286,6 +272,6 @@ class Calendar extends EntityRepository
         $qb->setParameter(100, $calendar);
 
 
-        return ! is_null($qb->getQuery()->getOneOrNullResult());
+        return null !== $qb->getQuery()->getOneOrNullResult();
     }
 }
