@@ -7,75 +7,84 @@
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
  * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
  */
+declare(strict_types=1);
+
 namespace Calendar\Service;
 
 use Calendar\Entity;
 use Calendar\Entity\Calendar;
 use Calendar\Entity\Contact as CalendarContact;
 use Calendar\Repository;
+use Calendar\Search\Service\CalendarSearchService;
 use Contact\Entity\Contact;
+use Contact\Service\ContactService;
+use Contact\Service\SelectionContactService;
+use Doctrine\ORM\EntityManager;
 use Project\Entity\Project;
+use Search\Service\AbstractSearchService;
+use Search\Service\SearchUpdateInterface;
+use Solarium\Client;
+use Solarium\Core\Query\AbstractQuery;
+use Solarium\QueryType\Update\Query\Document\Document;
 
 /**
+ * Class CalendarService
  *
+ * @package Calendar\Service
  */
-class CalendarService extends ServiceAbstract
+class CalendarService extends AbstractService implements SearchUpdateInterface
 {
-    /**
-     * Constant to determine which affiliations must be taken from the database
-     */
-    const WHICH_UPCOMING = 'upcoming';
-    const WHICH_UPDATED = 'updated';
-    const WHICH_PAST = 'past';
-    const WHICH_FINAL = 'final';
-    const WHICH_REVIEWS = 'project-reviews';
-    const WHICH_ON_HOMEPAGE = 'on-homepage';
-
+    public const WHICH_UPCOMING = 'Upcoming';
+    public const WHICH_UPDATED = 'Updated';
+    public const WHICH_PAST = 'Past';
+    public const WHICH_FINAL = 'Final';
+    public const WHICH_REVIEWS = 'Reviews';
+    public const WHICH_ON_HOMEPAGE = 'Homepage';
+    public const WHICH_HIGHLIGHT = 'Highlight';
 
     /**
-     * @param $id
-     *
-     * @return null|Calendar|object
+     * @var CalendarSearchService
      */
-    public function findCalendarById($id)
-    {
-        return $this->getEntityManager()->getRepository(Calendar::class)->find($id);
+    private $calendarSearchService;
+    /**
+     * @var ContactService
+     */
+    private $contactService;
+
+    public function __construct(
+        EntityManager $entityManager,
+        SelectionContactService $selectionContactService,
+        CalendarSearchService $calendarSearchService,
+        ContactService $contactService
+    ) {
+        parent::__construct($entityManager, $selectionContactService);
+
+        $this->calendarSearchService = $calendarSearchService;
+        $this->contactService = $contactService;
     }
 
-    /**
-     * @param $docRef
-     *
-     * @return null|Entity\Calendar|object
-     */
-    public function findCalendarByDocRef($docRef)
+    public function findCalendarById(int $id): ?Calendar
     {
-        return $this->getEntityManager()->getRepository(Entity\Calendar::class)->findOneBy(
+        return $this->entityManager->getRepository(Calendar::class)->find($id);
+    }
+
+    public function findCalendarByDocRef(string $docRef): ?Calendar
+    {
+        return $this->entityManager->getRepository(Entity\Calendar::class)->findOneBy(
             [
                 'docRef' => $docRef,
             ]
         );
     }
 
-    /**
-     * @param array    $data
-     * @param Calendar $calendar
-     *
-     * array (size=5)
-     * 'type' => string '2' (length=1)
-     * 'added' => string '20388' (length=5)
-     * 'removed' => string '' (length=0)
-     * 'sql' => string '' (length=0)
-     *
-     *
-     */
-    public function updateCalendarContacts(Calendar $calendar, array $data)
+    public function updateCalendarContacts(Calendar $calendar, array $data): void
     {
         //Update the contacts
-        if (! empty($data['added'])) {
+        if (!empty($data['added'])) {
             foreach (explode(',', $data['added']) as $contactId) {
-                $contact = $this->getContactService()->findEntityById(Contact::class, $contactId);
+                $contact = $this->contactService->findContactById((int)$contactId);
 
-                if (! is_null($contact) && ! $this->calendarHasContact($calendar, $contact)) {
+                if (null !== $contact && !$this->calendarHasContact($calendar, $contact)) {
                     $calendarContact = new CalendarContact();
                     $calendarContact->setContact($contact);
                     $calendarContact->setCalendar($calendar);
@@ -85,160 +94,233 @@ class CalendarService extends ServiceAbstract
                      *
                      * @var $role Entity\ContactRole
                      */
-                    $role = $this->findEntityById(Entity\ContactRole::class, Entity\ContactRole::ROLE_ATTENDEE);
+                    $role = $this->find(Entity\ContactRole::class, Entity\ContactRole::ROLE_ATTENDEE);
                     $calendarContact->setRole($role);
 
                     /**
-                     * Add every new user as attendee
+                     * Give every new user the status "tentative"
                      *
                      * @var $status Entity\ContactStatus
                      */
-                    $status = $this->findEntityById(
-                        Entity\ContactStatus::class,
-                        Entity\ContactStatus::STATUS_TENTATIVE
-                    );
+                    $status = $this->find(Entity\ContactStatus::class, Entity\ContactStatus::STATUS_TENTATIVE);
                     $calendarContact->setStatus($status);
 
-                    $this->newEntity($calendarContact);
+                    $this->save($calendarContact);
                 }
             }
         }
 
         //Update the contacts
-        if (! empty($data['removed'])) {
-            foreach (explode(',', $data['removed']) as $contactId) {
+        if (!empty($data['removed'])) {
+            foreach (\explode(',', $data['removed']) as $contactId) {
                 foreach ($calendar->getCalendarContact() as $calendarContact) {
                     if ($calendarContact->getContact()->getId() === (int)$contactId) {
-                        $this->removeEntity($calendarContact);
+                        $this->delete($calendarContact);
                     }
                 }
             }
         }
     }
 
-    /**
-     * @param Calendar $calendar
-     * @param Contact  $contact
-     *
-     * @return bool
-     */
-    public function calendarHasContact(Calendar $calendar, Contact $contact)
+    public function calendarHasContact(Calendar $calendar, Contact $contact): bool
     {
-        $calendarContact = $this->getEntityManager()->getRepository(CalendarContact::class)->findOneBy(
+        $calendarContact = $this->entityManager->getRepository(CalendarContact::class)->findOneBy(
             [
                 'calendar' => $calendar,
                 'contact'  => $contact,
             ]
         );
 
-        return ! is_null($calendarContact);
+        return null !== $calendarContact;
+    }
+
+    public function save(Entity\AbstractEntity $abstractEntity): Entity\AbstractEntity
+    {
+        parent::save($abstractEntity);
+
+        if ($abstractEntity instanceof Calendar) {
+            $this->updateEntityInSearchEngine($abstractEntity);
+        }
+
+        return $abstractEntity;
     }
 
     /**
-     * @param  string  $which
-     * @param  Contact $contact
-     *
-     * @return CalendarContact[]
+     * @param Calendar $calendar
      */
+    public function updateEntityInSearchEngine($calendar): void
+    {
+        $document = $this->prepareSearchUpdate($calendar);
+
+        $this->calendarSearchService->executeUpdateDocument($document);
+    }
+
+    /**
+     * @param Calendar $calendar
+     *
+     * @return AbstractQuery
+     */
+    public function prepareSearchUpdate($calendar): AbstractQuery
+    {
+        $searchClient = new Client();
+        $update = $searchClient->createUpdate();
+
+        /** @var Document $calendarDocument */
+        $calendarDocument = $update->createDocument();
+
+        // Calendar properties
+        $calendarDocument->setField('id', $calendar->getResourceId());
+        $calendarDocument->setField('calendar_id', $calendar->getId());
+
+        $calendarDocument->setField('docref', $calendar->getDocRef());
+
+        $calendarDocument->setField('calendar', $calendar->getCalendar());
+        $calendarDocument->setField('calendar_sort', $calendar->getCalendar());
+        $calendarDocument->setField('calendar_search', $calendar->getCalendar());
+
+        $calendarDocument->setField('description', $calendar->getDescription());
+        $calendarDocument->setField('description_sort', $calendar->getDescription());
+        $calendarDocument->setField('description_search', $calendar->getDescription());
+
+        if ($calendar->isHighlight()) {
+            $calendarDocument->setField('highlight_description', $calendar->getHighlightDescription());
+            $calendarDocument->setField('highlight_description_sort', $calendar->getHighlightDescription());
+            $calendarDocument->setField('highlight_description_search', $calendar->getHighlightDescription());
+        }
+
+        $calendarDocument->setField('location', $calendar->getLocation());
+        $calendarDocument->setField('location_sort', $calendar->getLocation());
+        $calendarDocument->setField('location_search', $calendar->getLocation());
+
+        if (null !== $calendar->getDateFrom()) {
+            $calendarDocument->setField(
+                'date_from',
+                $calendar->getDateFrom()->format(AbstractSearchService::DATE_SOLR)
+            );
+        }
+        if (null !== $calendar->getDateEnd()) {
+            $calendarDocument->setField('date_end', $calendar->getDateEnd()->format(AbstractSearchService::DATE_SOLR));
+        }
+        if (null !== $calendar->getDateCreated()) {
+            $calendarDocument->setField(
+                'date_created',
+                $calendar->getDateCreated()->format(AbstractSearchService::DATE_SOLR)
+            );
+        }
+        if (null !== $calendar->getDateUpdated()) {
+            $calendarDocument->setField(
+                'date_updated',
+                $calendar->getDateUpdated()->format(AbstractSearchService::DATE_SOLR)
+            );
+        }
+
+        $calendarDocument->setField('year', $calendar->getDateFrom()->format('Y'));
+        $calendarDocument->setField('month', $calendar->getDateFrom()->format('m'));
+
+        $calendarDocument->setField('highlight', $calendar->isHighlight());
+        $calendarDocument->setField('highlight_text', $calendar->isHighlight() ? 'Yes' : 'No');
+        $calendarDocument->setField('own_event', $calendar->isOwnEvent());
+        $calendarDocument->setField('own_event_text', $calendar->isOwnEvent() ? 'Yes' : 'No');
+        $calendarDocument->setField('is_present', $calendar->isPresent());
+        $calendarDocument->setField('is_present_text', $calendar->isPresent() ? 'Yes' : 'No');
+        $calendarDocument->setField('on_homepage', $calendar->onHomepage());
+        $calendarDocument->setField('on_homepage_text', $calendar->onHomepage() ? 'Yes' : 'No');
+
+        $update->addDocument($calendarDocument);
+        $update->addCommit();
+
+        return $update;
+    }
+
+    public function delete(Entity\AbstractEntity $abstractEntity): void
+    {
+        if ($abstractEntity instanceof Calendar) {
+            $this->calendarSearchService->deleteDocument($abstractEntity);
+        }
+
+        parent::delete($abstractEntity);
+    }
+
     public function findCalendarContactByContact(
-        $which = self::WHICH_UPCOMING,
-        Contact $contact = null
-    ) {
+        string $which,
+        Contact $contact
+    ): array {
         /** @var Repository\Contact $repository */
-        $repository = $this->getEntityManager()->getRepository(CalendarContact::class);
+        $repository = $this->entityManager->getRepository(CalendarContact::class);
 
         return $repository->findCalendarContactByContact($which, $contact);
     }
 
-    /**
-     * @param Contact  $contact
-     * @param Calendar $calendar
-     *
-     * @return CalendarContact
-     */
     public function findCalendarContactByContactAndCalendar(
         Contact $contact,
         Calendar $calendar
-    ) {
+    ): ?CalendarContact {
         /** @var Repository\Contact $repository */
-        $repository = $this->getEntityManager()->getRepository(CalendarContact::class);
+        $repository = $this->entityManager->getRepository(CalendarContact::class);
 
         return $repository->findCalendarContactByContactAndCalendar($contact, $calendar);
     }
 
-    /**
-     * @param Calendar $calendar
-     * @param int      $status
-     *
-     * @return CalendarContact[]
-     */
-    public function findCalendarContactsByCalendar(Calendar $calendar, $status = CalendarContact::STATUS_ALL)
-    {
+    public function findCalendarContactsByCalendar(
+        Calendar $calendar,
+        $status = CalendarContact::STATUS_ALL,
+        string $order = 'lastname'
+    ): array {
         /** @var Repository\Contact $repository */
-        $repository = $this->getEntityManager()->getRepository(CalendarContact::class);
+        $repository = $this->entityManager->getRepository(CalendarContact::class);
 
-        return $repository->findCalendarContactsByCalendar($calendar, $status);
+        return $repository->findCalendarContactsByCalendar($calendar, $status, $order);
     }
 
-    /**
-     * This function will return a boolean value to see if a contact can view the calendar
-     *
-     * @param Calendar $calendar
-     * @param Contact  $contact
-     *
-     * @return bool
-     */
-    public function canViewCalendar(Calendar $calendar, Contact $contact = null)
+    public function canViewCalendar(Calendar $calendar, Contact $contact = null): bool
     {
         /** @var Repository\Calendar $repository */
-        $repository = $this->getEntityManager()->getRepository(Entity\Calendar::class);
+        $repository = $this->entityManager->getRepository(Entity\Calendar::class);
 
         return $repository->canViewCalendar($calendar, $contact);
     }
 
-    /**
-     * @param string  $which
-     * @param Contact $contact
-     * @param integer $year
-     * @param integer $type
-     *
-     * @return \Doctrine\ORM\Query
-     */
     public function findCalendarItems(
-        $which = self::WHICH_UPCOMING,
+        string $which = self::WHICH_UPCOMING,
         Contact $contact = null,
-        $year = null,
-        $type = null
-    ) {
+        int $year = null,
+        string $type = null
+    ): \Doctrine\ORM\Query {
         /** @var \Calendar\Repository\Calendar $repository */
-        $repository = $this->getEntityManager()->getRepository(Entity\Calendar::class);
+        $repository = $this->entityManager->getRepository(Entity\Calendar::class);
 
-        return $repository->findCalendarItems($which, true, $contact, $year, $type);
+        $limitQueryBuilder = null;
+        if (null !== $contact) {
+            /*
+             * Grab the limiting query-builder from the AdminService
+             */
+            $limitQueryBuilder = $this->parseWherePermit(
+                new Calendar(),
+                'view',
+                $contact
+            );
+        }
+
+        return $repository->findCalendarItems($which, true, $contact, $year, $type, $limitQueryBuilder);
     }
 
-    /**
-     * @param bool    $onlyFinal
-     * @param Project $project
-     *
-     * @return Calendar[]
-     */
-    public function findCalendarByProject(Project $project, $onlyFinal = true)
+    public function findCalendarByProject(Project $project, $onlyFinal = true): array
     {
         $calendar = [];
         /**
          * Add the calendar items from the project
          */
         foreach ($project->getProjectCalendar() as $calendarItem) {
-            if (! $onlyFinal
-                 || $calendarItem->getCalendar()->getFinal() === Calendar::FINAL_FINAL
+            if (!$onlyFinal
+                || $calendarItem->getCalendar()->getFinal() === Calendar::FINAL_FINAL
             ) {
                 $calendar[$calendarItem->getCalendar()->getId()]
                     = $calendarItem->getCalendar();
             }
         }
         foreach ($project->getCall()->getCalendar() as $calendarItem) {
-            if (! $onlyFinal
-                 || $calendarItem->getFinal() === Calendar::FINAL_FINAL
+            if (!$onlyFinal
+                || $calendarItem->getFinal() === Calendar::FINAL_FINAL
             ) {
                 if ($calendarItem->getDateEnd() > new \DateTime()) {
                     $calendar[$calendarItem->getId()] = $calendarItem;
@@ -249,104 +331,65 @@ class CalendarService extends ServiceAbstract
         return $calendar;
     }
 
-    /**
-     * return the review-meeting corresponding to a calendar item
-     *
-     * @param Project $project
-     *
-     * @return Calendar|null
-     */
-    public function findLatestProjectCalendar(Project $project)
+    public function findLatestProjectCalendar(Project $project): ?Calendar
     {
         /** @var \Calendar\Repository\Calendar $repository */
-        $repository = $this->getEntityManager()->getRepository(Entity\Calendar::class);
+        $repository = $this->entityManager->getRepository(Entity\Calendar::class);
 
         return $repository->findLatestProjectCalendar($project);
     }
 
-    /**
-     * Return the news review meeting
-     *
-     * @param Project   $project
-     * @param \DateTime $datetime
-     *
-     * @return Calendar|null
-     */
     public function findNextProjectCalendar(
         Project $project,
         \DateTime $datetime
-    ) {
+    ): ?Calendar {
         /** @var \Calendar\Repository\Calendar $repository */
-        $repository = $this->getEntityManager()->getRepository(Entity\Calendar::class);
+        $repository = $this->entityManager->getRepository(Entity\Calendar::class);
 
         return $repository->findNextProjectCalendar($project, $datetime);
     }
 
-    /**
-     * Return the lastest review meeting
-     *
-     * @param Project   $project
-     * @param \DateTime $datetime
-     *
-     * @return Calendar|null
-     */
     public function findPreviousProjectCalendar(
         Project $project,
         \DateTime $datetime
-    ) {
+    ): ?Calendar {
         /** @var \Calendar\Repository\Calendar $repository */
-        $repository = $this->getEntityManager()->getRepository(Entity\Calendar::class);
+        $repository = $this->entityManager->getRepository(Entity\Calendar::class);
 
         return $repository->findPreviousProjectCalendar($project, $datetime);
     }
 
-    /**
-     * @param CalendarContact $calendarContact
-     * @param                 $status
-     */
     public function updateContactStatus(
         CalendarContact $calendarContact,
-        $status
-    ) {
-        $calendarContact->setStatus($this->getEntityManager()->getReference(Entity\ContactStatus::class, $status));
-        $this->updateEntity($calendarContact);
+        string $status
+    ): void {
+        /** @var Entity\ContactStatus $contactStatus */
+        $contactStatus = $this->entityManager->getReference(Entity\ContactStatus::class, $status);
+
+        $calendarContact->setStatus($contactStatus);
+        $this->save($calendarContact);
     }
 
-    /**
-     * @param Calendar $calendar
-     *
-     * @return CalendarContact[]
-     */
-    public function findGeneralCalendarContactByCalendar(Calendar $calendar)
+    public function findGeneralCalendarContactByCalendar(Calendar $calendar): array
     {
         /** @var Repository\Contact $repository */
-        $repository = $this->getEntityManager()->getRepository(CalendarContact::class);
+        $repository = $this->entityManager->getRepository(CalendarContact::class);
 
         return $repository->findGeneralCalendarContactByCalendar($calendar);
     }
 
-    /**
-     * @return \stdClass
-     */
-    public function findMinAndMaxYear()
+    public function updateCollectionInSearchEngine(bool $clearIndex = false): void
     {
-        /** @var Repository\Calendar $repository */
-        $repository = $this->getEntityManager()->getRepository(Entity\Calendar::class);
+        $calendarItems = $this->findAll(Entity\Calendar::class);
+        $collection = [];
+        foreach ($calendarItems as $calendar) {
+            $collection[] = $this->prepareSearchUpdate($calendar);
+        }
 
-        $yearSpanResult    = $repository->findMinAndMaxYear();
-        $yearSpan          = new \stdClass();
-        $yearSpan->minYear = (int)$yearSpanResult['minYear'];
-        $yearSpan->maxYear = (int)$yearSpanResult['maxYear'];
-
-        return $yearSpan;
+        $this->calendarSearchService->updateIndexWithCollection($collection, $clearIndex);
     }
 
-    /**
-     * Return an array of all which-values
-     *
-     * @return array
-     */
-    public function getWhichValues()
+    public function getWhichValues(): array
     {
         return [
             self::WHICH_UPCOMING,
@@ -354,6 +397,7 @@ class CalendarService extends ServiceAbstract
             self::WHICH_FINAL,
             self::WHICH_PAST,
             self::WHICH_REVIEWS,
+            self::WHICH_ON_HOMEPAGE
         ];
     }
 }
