@@ -8,6 +8,8 @@
  * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
  */
 
+declare(strict_types=1);
+
 namespace Calendar\Version;
 
 /**
@@ -15,8 +17,5 @@ namespace Calendar\Version;
  */
 final class Version
 {
-    /**
-     * Zend Framework version identification - see compareVersion().
-     */
-    const VERSION = '3.0.0';
+    public const VERSION = '4.0.0';
 }
