@@ -13,6 +13,8 @@
  * @link        http://github.com/iteaoffice/project for the canonical source repository
  */
 
+declare(strict_types=1);
+
 namespace Calendar\View\Factory;
 
 use Calendar\View\Helper\AbstractViewHelper;
@@ -32,8 +34,8 @@ final class ViewHelperFactory implements FactoryInterface
      * Create an instance of the requested class name.
      *
      * @param ContainerInterface|HelperPluginManager $container
-     * @param string                                 $requestedName
-     * @param null|array                             $options
+     * @param string $requestedName
+     * @param null|array $options
      *
      * @return object
      */
